core_modules/domain_ip_analysis/requirements.txt
python-whois
dnspython
shodan
click
